Move selector out of section
This is required due to the processing of the selector causing issues with strings passed in with replaced portions in them. More specifically, passing `Admin user {{userEntity.username}} created.` from a test prevents the selector from processing, passing the selector function string (`{{AdminEventLogDigestSection.messageWithUser(...)}}`) to the XPath parser instead of the real selector.